fix post-create.sh

+++ .devcontainer/post-create.sh
@@ -11,24 +11,30 @@ fi
 # ln -s /workspaces/${PLUGIN_NAME} plugins/${PLUGIN_NAME}
 if [ -f plugins/${PLUGIN_NAME}/Gemfile_for_test ]
 then
-    cp plugins/${PLUGIN_NAME}/Gemfile_for_test plugins/${PLUGIN_NAME}/Gemfile 
+    cp plugins/${PLUGIN_NAME}/Gemfile_for_test plugins/${PLUGIN_NAME}/Gemfile
 fi
 
 
-bundle install 
+bundle install
+bundle exec rake redmine:plugins:ai_helper:setup_scm
 
 initdb() {
-    bundle exec rake db:create
+    if [ $DB != "sqlite3" ]
+    then
+        bundle exec rake db:create
+    fi
     bundle exec rake db:migrate
     bundle exec rake redmine:plugins:migrate
 
-    bundle exec rake db:drop RAILS_ENV=test
-    bundle exec rake db:create RAILS_ENV=test
+    if [ $DB != "sqlite3" ]
+    then
+        bundle exec rake db:drop RAILS_ENV=test
+        bundle exec rake db:create RAILS_ENV=test
+    fi
+
     bundle exec rake db:migrate RAILS_ENV=test
     bundle exec rake redmine:plugins:migrate RAILS_ENV=test
 }
-
-initdb
 
 export DB=mysql2
 export DB_NAME=redmine
@@ -46,4 +52,8 @@ export DB_PASSWORD=postgres
 export DB_HOST=postgres
 export DB_PORT=5432
 
+initdb
+
+rm -f db/redmine.sqlite3_test
+export DB=sqlite3
 initdb
